requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Pillow==10.4.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
bcrypt==4.0.1